+++ contracts/interfaces/IAuction.sol
@@ -3,10 +3,10 @@ pragma solidity ^0.8.9;
 
 interface IAuction {
     struct AuctionData {
-        uint startTime;
-        uint currentBid;
+        uint256 startTime;
+        uint256 currentBid;
         address currentBidder;
         bool finalized;
-        uint tokenAmount;
+        uint256 tokenAmount;
     }
 }
