+++ contracts/interfaces/IBuyout.sol
@@ -3,10 +3,10 @@ pragma solidity ^0.8.9;
 
 interface IBuyout {
     struct Buyout {
-        uint startTime;
-        uint snapshotId;
-        uint competingPoolBalance;
-        uint originalPoolBalance;
+        uint256 startTime;
+        uint256 snapshotId;
+        uint256 competingPoolBalance;
+        uint256 originalPoolBalance;
         address transferTarget;
     }
 }
