+++ contracts/interfaces/IGrief.sol
@@ -5,10 +5,10 @@ interface IGrief {
     struct VoteData {
         Invocation invocation;
         VoteOutcome outcome;
-        uint snapshotId;
-        uint startTime;
-        uint inflationTotal;
-        uint remainingTokenAmount;
+        uint256 snapshotId;
+        uint256 startTime;
+        uint256 inflationTotal;
+        uint256 remainingTokenAmount;
         bool isGrief;
         bool finalized;
         bool merged;
