+++ contracts/interfaces/ISPOG.sol
@@ -4,7 +4,9 @@ pragma solidity ^0.8.9;
 import {IGrief} from "./IGrief.sol";
 
 interface ISPOG is IGrief {
-    function stake(uint amount) external;
-    function claim(uint voteId) external;
-    function vote(uint voteId, bool affirmative) external;
+    function stake(uint256 amount) external;
+
+    function claim(uint256 voteId) external;
+
+    function vote(uint256 voteId, bool affirmative) external;
 }
